Extracts placement code.
