+++ oatie/src/lib.rs
@@ -55,7 +55,8 @@ pub mod macros;
 mod parse;
 pub mod schema;
 pub mod stepper;
-pub mod string;
+mod string;
+mod place;
 pub mod transform;
 pub mod transform_test;
 pub mod validate;
